requirements.txt
Flask==3.0.0
requests==2.31.0
httpx==0.27.2
openai==1.55.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
